Distribute deserialization tests over command modules
