+++ epp-client/src/contact/check.rs
@@ -1,10 +1,9 @@
 use std::fmt::Debug;
 
 /// Types for EPP contact check request
-
 use super::XMLNS;
 use crate::common::{NoExtension, StringValue};
-use crate::request::{Transaction, Command};
+use crate::request::{Command, Transaction};
 use serde::{Deserialize, Serialize};
 
 impl Transaction<NoExtension> for ContactCheck {}
@@ -88,4 +87,34 @@ pub struct ContactCheckResponse {
     /// Data under the &lt;chkData&gt; tag
     #[serde(rename = "chkData")]
     pub check_data: ContactCheckResponseData,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::ContactCheck;
+    use crate::request::Transaction;
+    use crate::tests::{get_xml, CLTRID, SUCCESS_MSG, SVTRID};
+
+    #[test]
+    fn contact_check() {
+        let xml = get_xml("response/contact/check.xml").unwrap();
+        let object = ContactCheck::deserialize_response(xml.as_str()).unwrap();
+
+        let results = object.res_data().unwrap();
+
+        assert_eq!(object.result.code, 1000);
+        assert_eq!(object.result.message, SUCCESS_MSG.into());
+        assert_eq!(
+            results.check_data.contact_list[0].contact.id,
+            "eppdev-contact-1".into()
+        );
+        assert_eq!(results.check_data.contact_list[0].contact.available, 0);
+        assert_eq!(
+            results.check_data.contact_list[1].contact.id,
+            "eppdev-contact-2".into()
+        );
+        assert_eq!(results.check_data.contact_list[1].contact.available, 1);
+        assert_eq!(object.tr_ids.client_tr_id.unwrap(), CLTRID.into());
+        assert_eq!(object.tr_ids.server_tr_id, SVTRID.into());
+    }
 }
