+++ epp-client/src/contact/create.rs
@@ -2,7 +2,7 @@
 
 use super::XMLNS;
 use crate::common::{ContactAuthInfo, NoExtension, Phone, PostalInfo, StringValue};
-use crate::request::{Transaction, Command};
+use crate::request::{Command, Transaction};
 use serde::{Deserialize, Serialize};
 
 impl Transaction<NoExtension> for ContactCreate {}
@@ -96,4 +96,29 @@ pub struct ContactCreateResponse {
     /// Data under the &lt;creData&gt; tag
     #[serde(rename = "creData")]
     pub create_data: ContactCreateData,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::ContactCreate;
+    use crate::request::Transaction;
+    use crate::tests::{get_xml, CLTRID, SUCCESS_MSG, SVTRID};
+
+    #[test]
+    fn contact_create() {
+        let xml = get_xml("response/contact/create.xml").unwrap();
+        let object = ContactCreate::deserialize_response(xml.as_str()).unwrap();
+
+        let results = object.res_data().unwrap();
+
+        assert_eq!(object.result.code, 1000);
+        assert_eq!(object.result.message, SUCCESS_MSG.into());
+        assert_eq!(results.create_data.id, "eppdev-contact-4".into());
+        assert_eq!(
+            results.create_data.created_at,
+            "2021-07-25T16:05:32.0Z".into()
+        );
+        assert_eq!(object.tr_ids.client_tr_id.unwrap(), CLTRID.into());
+        assert_eq!(object.tr_ids.server_tr_id, SVTRID.into());
+    }
 }
