+++ epp-client/src/contact/delete.rs
@@ -41,3 +41,21 @@ impl ContactDelete {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::ContactDelete;
+    use crate::request::Transaction;
+    use crate::tests::{get_xml, CLTRID, SUCCESS_MSG, SVTRID};
+
+    #[test]
+    fn contact_delete() {
+        let xml = get_xml("response/contact/delete.xml").unwrap();
+        let object = ContactDelete::deserialize_response(xml.as_str()).unwrap();
+
+        assert_eq!(object.result.code, 1000);
+        assert_eq!(object.result.message, SUCCESS_MSG.into());
+        assert_eq!(object.tr_ids.client_tr_id.unwrap(), CLTRID.into());
+        assert_eq!(object.tr_ids.server_tr_id, SVTRID.into());
+    }
+}
