+++ epp-client/src/domain/info.rs
@@ -132,3 +132,75 @@ pub struct DomainInfoResponse {
     #[serde(rename = "infData")]
     pub info_data: DomainInfoResponseData,
 }
+
+#[cfg(test)]
+mod tests {
+    use super::DomainInfo;
+    use crate::common::NoExtension;
+    use crate::request::Transaction;
+    use crate::tests::{get_xml, CLTRID, SUCCESS_MSG, SVTRID};
+
+    #[test]
+    fn domain_info() {
+        let xml = get_xml("response/domain/info.xml").unwrap();
+        let object =
+            <DomainInfo as Transaction<NoExtension>>::deserialize_response(xml.as_str()).unwrap();
+
+        let result = object.res_data().unwrap();
+        let auth_info = result.info_data.auth_info.as_ref().unwrap();
+        let ns_list = result.info_data.ns.as_ref().unwrap();
+        let ns = (*ns_list).host_obj.as_ref().unwrap();
+        let hosts = result.info_data.hosts.as_ref().unwrap();
+        let statuses = result.info_data.statuses.as_ref().unwrap();
+        let registrant = result.info_data.registrant.as_ref().unwrap();
+        let contacts = result.info_data.contacts.as_ref().unwrap();
+
+        assert_eq!(object.result.code, 1000);
+        assert_eq!(object.result.message, SUCCESS_MSG.into());
+        assert_eq!(result.info_data.name, "eppdev-1.com".into());
+        assert_eq!(result.info_data.roid, "125899511_DOMAIN_COM-VRSN".into());
+        assert_eq!(statuses[0].status, "ok".to_string());
+        assert_eq!(statuses[1].status, "clientTransferProhibited".to_string());
+        assert_eq!(*registrant, "eppdev-contact-2".into());
+        assert_eq!(contacts[0].id, "eppdev-contact-2".to_string());
+        assert_eq!(contacts[0].contact_type, "admin".to_string());
+        assert_eq!(contacts[1].id, "eppdev-contact-2".to_string());
+        assert_eq!(contacts[1].contact_type, "tech".to_string());
+        assert_eq!(contacts[2].id, "eppdev-contact-2".to_string());
+        assert_eq!(contacts[2].contact_type, "billing".to_string());
+        assert_eq!((*ns)[0], "ns1.eppdev-1.com".into());
+        assert_eq!((*ns)[1], "ns2.eppdev-1.com".into());
+        assert_eq!((*hosts)[0], "ns1.eppdev-1.com".into());
+        assert_eq!((*hosts)[1], "ns2.eppdev-1.com".into());
+        assert_eq!(result.info_data.client_id, "eppdev".into());
+        assert_eq!(
+            *result.info_data.creator_id.as_ref().unwrap(),
+            "SYSTEM".into()
+        );
+        assert_eq!(
+            *result.info_data.created_at.as_ref().unwrap(),
+            "2021-07-23T15:31:20.0Z".into()
+        );
+        assert_eq!(
+            *result.info_data.updater_id.as_ref().unwrap(),
+            "SYSTEM".into()
+        );
+        assert_eq!(
+            *result.info_data.updated_at.as_ref().unwrap(),
+            "2021-07-23T15:31:21.0Z".into()
+        );
+        assert_eq!(
+            *result.info_data.expiring_at.as_ref().unwrap(),
+            "2023-07-23T15:31:20.0Z".into()
+        );
+        assert_eq!((*auth_info).password, "epP4uthd#v".into());
+        assert_eq!(object.tr_ids.client_tr_id.unwrap(), CLTRID.into());
+        assert_eq!(object.tr_ids.server_tr_id, SVTRID.into());
+    }
+
+    #[test]
+    fn domain_info_alt() {
+        let xml = get_xml("response/domain/info_alt.xml").unwrap();
+        <DomainInfo as Transaction<NoExtension>>::deserialize_response(xml.as_str()).unwrap();
+    }
+}
