+++ epp-client/src/extensions/rgp/request.rs
@@ -66,3 +66,43 @@ pub struct RgpRequestResponse {
     #[serde(rename = "rgpStatus")]
     pub rgp_status: Vec<RgpStatus>,
 }
+
+#[cfg(test)]
+mod tests {
+    use super::{RgpRestoreRequest, Update};
+    use crate::domain::info::DomainInfo;
+    use crate::domain::update::DomainUpdate;
+    use crate::request::Transaction;
+    use crate::tests::{get_xml, SUCCESS_MSG, SVTRID};
+
+    #[test]
+    fn rgp_restore_request() {
+        let xml = get_xml("response/extensions/rgp_restore.xml").unwrap();
+        let object =
+            <DomainUpdate as Transaction<Update<RgpRestoreRequest>>>::deserialize_response(
+                xml.as_str(),
+            )
+            .unwrap();
+
+        let ext = object.extension.unwrap();
+
+        assert_eq!(object.result.code, 1000);
+        assert_eq!(object.result.message, SUCCESS_MSG.into());
+        assert_eq!(ext.data.rgp_status[0].status, "pendingRestore".to_string());
+        assert_eq!(object.tr_ids.server_tr_id, SVTRID.into());
+    }
+
+    #[test]
+    fn rgp_restore_domain_info_response() {
+        let xml = get_xml("response/extensions/domain_info_rgp.xml").unwrap();
+        let object = <DomainInfo as Transaction<Update<RgpRestoreRequest>>>::deserialize_response(
+            xml.as_str(),
+        )
+        .unwrap();
+
+        let ext = object.extension.unwrap();
+
+        assert_eq!(ext.data.rgp_status[0].status, "addPeriod");
+        assert_eq!(ext.data.rgp_status[1].status, "renewPeriod");
+    }
+}
