+++ epp-client/src/host/update.rs
@@ -87,3 +87,21 @@ pub struct HostUpdate {
     #[serde(rename = "host:update", alias = "update")]
     host: HostUpdateRequestData,
 }
+
+#[cfg(test)]
+mod tests {
+    use super::HostUpdate;
+    use crate::request::Transaction;
+    use crate::tests::{get_xml, CLTRID, SUCCESS_MSG, SVTRID};
+
+    #[test]
+    fn host_update() {
+        let xml = get_xml("response/host/update.xml").unwrap();
+        let object = HostUpdate::deserialize_response(xml.as_str()).unwrap();
+
+        assert_eq!(object.result.code, 1000);
+        assert_eq!(object.result.message, SUCCESS_MSG.into());
+        assert_eq!(object.tr_ids.client_tr_id.unwrap(), CLTRID.into());
+        assert_eq!(object.tr_ids.server_tr_id, SVTRID.into());
+    }
+}
