+++ epp-client/src/tests/mod.rs
@@ -1,16 +1,17 @@
 //! Module for automated tests
 
-pub mod de;
 pub mod se;
 
 use regex::Regex;
 use std::{error::Error, fs::File, io::Read};
 
-const RESOURCES_DIR: &str = "./test/resources";
-const CLTRID: &str = "cltrid:1626454866";
+pub(crate) const RESOURCES_DIR: &str = "./test/resources";
+pub(crate) const CLTRID: &str = "cltrid:1626454866";
+pub(crate) const SVTRID: &str = "RO-6879-1627224678242975";
+pub(crate) const SUCCESS_MSG: &str = "Command completed successfully";
 
 /// Reads EPP XML requests and responses from the test/resources directory to run tests on
-fn get_xml(path: &str) -> Result<String, Box<dyn Error>> {
+pub(crate) fn get_xml(path: &str) -> Result<String, Box<dyn Error>> {
     let ws_regex = Regex::new(r"[\s]{2,}")?;
 
     let mut f = File::open(format!("{}/{}", RESOURCES_DIR, path))?;
